+++ x-pack/plugins/monitoring/server/lib/metrics/classes/cluster_metric.js
@@ -1,0 +1,16 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+import { Metric } from './metric';
+
+export class ClusterMetric extends Metric {
+  constructor(opts) {
+    super({
+      ...opts,
+      uuidField: 'cluster_uuid'
+    });
+  }
+}
