+++ x-pack/plugins/monitoring/server/lib/metrics/classes/index.js
@@ -1,0 +1,9 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export { Metric } from './metric';
+export { ClusterMetric } from './cluster_metric';
+export { QuotaMetric } from './quota_metric';
