+++ x-pack/plugins/monitoring/server/lib/metrics/kibana/classes.js
@@ -1,0 +1,81 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+import { ClusterMetric, Metric  } from '../classes';
+import { NORMALIZED_DERIVATIVE_UNIT } from '../../../../common/constants';
+
+export class KibanaClusterMetric extends ClusterMetric {
+  constructor(opts) {
+    super({
+      ...opts,
+      app: 'kibana',
+      ...KibanaClusterMetric.getMetricFields()
+    });
+  }
+
+  static getMetricFields() {
+    return {
+      uuidField: 'cluster_uuid',
+      timestampField: 'kibana_stats.timestamp'
+    };
+  }
+}
+
+export class KibanaEventsRateClusterMetric extends KibanaClusterMetric {
+  constructor(opts) {
+    super({
+      ...opts,
+      metricAgg: 'max'
+    });
+
+    this.aggs = {
+      kibana_uuids: {
+        terms: {
+          field: 'kibana_stats.kibana.uuid',
+          size: 1000
+        },
+        aggs: {
+          event_rate_per_instance: {
+            max: {
+              field: this.field
+            }
+          }
+        }
+      },
+      event_rate: {
+        sum_bucket: {
+          buckets_path: 'kibana_uuids>event_rate_per_instance',
+          gap_policy: 'skip'
+        }
+      },
+      metric_deriv: {
+        derivative: {
+          buckets_path: 'event_rate',
+          gap_policy: 'skip',
+          unit: NORMALIZED_DERIVATIVE_UNIT
+        }
+      }
+    };
+  }
+}
+
+export class KibanaMetric extends Metric {
+  constructor(opts) {
+    super({
+      ...opts,
+      app: 'kibana',
+      ...KibanaMetric.getMetricFields()
+    });
+  }
+
+  static getMetricFields() {
+    return {
+      uuidField: 'kibana_stats.kibana.uuid',
+      timestampField: 'kibana_stats.timestamp'
+    };
+  }
+}
+
